pyproject.toml
[build-system]
requires = ["setuptools>=65", "setuptools_scm[toml]>=6.2", "wheel"]
# requires = ["setuptools>=65", "setuptools-git-versioning<2", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
# write_to = "wrfplot/_version.py"
version_scheme = "post-release"
# version_scheme = "python-simplified-semver"

[project]
name = "wrfplot"
description = "Command line application to plot WRF model output data"
dynamic = ["version",]
authors = [{ name = "J Sundar", email = "wrf.guy@gmail.com" }]
readme = "README.md"
requires-python = ">=3.7"
keywords = ["WRF", "Weather Research and Forecast", "Scientific", "Engineering", "Atmospheric Science", "Weather Model", "Plotting", "Software Development", "Numerical Weather Prediction", "NWP"]
license = {text = "GPL3"}
classifiers = [
	"License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: Unix",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: Implementation :: CPython",
]

dependencies = ['cartopy', 'xarray', 'matplotlib', 'wrf-python>=1.3', 'tqdm', 'netcdf4', 'colormaps']

[project.entry-points."wrfplot.main"]
wrfplot = "wrfplot:main"

[project.urls]
"Bug Reports" = "https://github.com/wxguy/wrfplot/issues"
"Source" = "https://github.com/wxguy/wrfplot/"
"Homepage" = "https://github.com/wxguy/wrfplot/"
"Documentation" = "https://wrfplot.readthedocs.io"

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages]
# find = {}

[tool.setuptools.packages.find]
namespaces = false
exclude = ["docs*", "tests*"    ]

[tool.setuptools.exclude-package-data]
"wrfplot" = [".gitattributes", "dist", "installer.iss", "tests", "docs", "installer.nsi", "environment.yml", "installer.sh", "release_check_list.py",
"wrfplot.egg-info", "dev", "install_env.txt", " __pycache__"]

[tool.pyright]
venvPath = "$HOME/miniconda3/envs"
venv = "wrfplot"
